requirements.txt
# PyQt6 依赖
PyQt6>=6.8.0          # GUI框架
PyQt6-sip>=13.9.0     # Python与Qt之间的SIP绑定

# 网络请求和API
requests>=2.31.0       # 用于API调用

# 环境配置
python-dotenv>=1.0.0   # 从.env文件加载环境变量

# 文件操作
zipfile36>=0.1.3       # 处理ZIP文件（测试数据打包） 